requirements.txt
# ===========================================
# Projeto Reforma/Eagle - Dependências
# ===========================================

# Django e extensões
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2

# Banco de Dados
psycopg2-binary>=2.9
dj-database-url>=2.1

# Celery e Redis (tarefas assíncronas)
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-celery-results>=2.5

# Servidor
gunicorn>=21.2
whitenoise>=6.6
uvicorn>=0.27

# Autenticação
django-allauth>=0.60

# Manipulação de Arquivos
Pillow>=10.2
openpyxl>=3.1
xlsxwriter>=3.1
lxml>=5.1
python-docx>=1.1
reportlab>=4.1

# HTTP e APIs
requests>=2.31
httpx>=0.26
aiohttp>=3.9

# Criptografia (certificados digitais)
cryptography>=42.0
pyOpenSSL>=24.0

# Configuração
python-decouple>=3.8
python-dotenv>=1.0

# XML e Parsing
xmltodict>=0.13
defusedxml>=0.7

# Validações
validate-docbr>=1.10

# Túnel para desenvolvimento
pyngrok>=7.0

# Utilitários
python-dateutil>=2.8
pytz>=2024.1

# Desenvolvimento
django-debug-toolbar>=4.2
ipython>=8.21

# Testes
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
coverage>=7.4
